==> 02-Ajustando Dataset.ipynb <==
from pathlib import Path
import pandas as pd

path_ds_training = Path("./OIA-ODIR/Training Set/Annotation/training annotation (English).xlsx")
path_ds_training_img = Path("./OIA-ODIR/Training Set/Images")
path_ds_validation = Path("./OIA-ODIR/Off-site Test Set/Annotation/off-site test annotation (English).xlsx")
path_ds_validation_img = Path("./OIA-ODIR/Off-site Test Set/Images")
path_ds_test = Path("./OIA-ODIR/On-site Test Set/Annotation/on-site test annotation (English).xlsx")
path_ds_test_img = Path("./OIA-ODIR/On-site Test Set/Images")

ds_training = pd.read_excel(path_ds_training, index_col=0)
ds_validation = pd.read_excel(path_ds_validation, index_col=0)
ds_test = pd.read_excel(path_ds_test, index_col=0)

### Ajustando o nome das colunas

columns = ["age", "sex", "left_file", "right_file", "left_keywords", "right_keywords", "N", "D", "G", "C", "A","H", "M", "O" ]
ds_training.columns = columns
ds_validation.columns = columns
ds_test.columns = columns

#### Excluindo registro com imagens ruins

# removendo imagens sem qualidade
to_drop = ds_training.query("left_file in (\"2174_left.jpg\", \"2175_left.jpg\", \"2176_left.jpg\", \"2177_left.jpg\", \"2178_left.jpg\", \"2180_left.jpg\", \"2181_left.jpg\", \"2182_left.jpg\", \"2957_left.jpg\")")
ds_training.drop(to_drop.index, inplace = True)

to_drop = ds_validation.query("left_file in (\"2174_left.jpg\", \"2175_left.jpg\", \"2176_left.jpg\", \"2177_left.jpg\", \"2178_left.jpg\", \"2180_left.jpg\", \"2181_left.jpg\", \"2182_left.jpg\", \"2957_left.jpg\")")
ds_validation.drop(to_drop.index, inplace = True)

to_drop = ds_test.query("left_file in (\"2174_left.jpg\", \"2175_left.jpg\", \"2176_left.jpg\", \"2177_left.jpg\", \"2178_left.jpg\", \"2180_left.jpg\", \"2181_left.jpg\", \"2182_left.jpg\", \"2957_left.jpg\")")
ds_test.drop(to_drop.index, inplace = True)


# Ajustando o dataset por olho

Na exporação do dataset descrobrimos que as classificações não estão por imagens e sim por Paciente, então vamos reorganizar o dataset por imagem e reclassifica-las
Depois vamos buscar imagens que foram diagnosticadas apenas com 1 doença e vamos desconsiderar imagens com mais de 1 classificação. Assim buscamos melhorar a performance da Rede.


### Metodo auxiliar para separar cada olho em uma linha

def get_ds_eyes_line(ds):
    left_features = ["left_file", "left_keywords" ]
    left_ds_eye = ds[left_features]
    left_ds_eye.insert(1, "side","left")

    left_ds_eye = left_ds_eye.rename(columns={
                                                "left_file": "file_name", 
                                                "left_keywords":"keywords"  })


    right_features = ["right_file", "right_keywords"]
    right_ds_eye = ds[right_features]
    right_ds_eye.insert(1, "side","right")


    right_ds_eye = right_ds_eye.rename(columns={"right_file": "file_name", 
                                                            "right_keywords":"keywords"})

    ds_eye = pd.concat([left_ds_eye, right_ds_eye], ignore_index=True)
    ds_eye =  ds_eye.sort_values(by=["file_name"],ignore_index=True)
    return ds_eye


ds_training_eye = get_ds_eyes_line(ds_training)
ds_validation_eye = get_ds_eyes_line(ds_validation)
ds_test_eye = get_ds_eyes_line(ds_test)


ds_training_eye.head()

### Metodos auxiliares para a reclassificação das imagens

# Metodos criados para reclassificar os imagens do Dataset

separador = ","

def ajusta_separador(keywords):
    return keywords.replace("，", ",").replace(",",",")

def classifica_desconsiderar(keywords):
    
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        #key == "lens dust" or key == "optic disk photographically invisible"
        if key == "low image quality" or key == "image offset" or key == "mild nonproliferative retinopathy" or key == "mild non proliferative retinopathy":
            return 1
    
    return 0

def classifica_normal(keywords):
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        if key == "normal fundus":
            return 1
    
    return 0

def classifica_diabetes(keywords):
    keys = keywords.split(separador)

    for key in keys:
        key = key.strip()
                   
        if key == "moderate non proliferative retinopathy" or key == "moderate nonproliferative retinopathy" or key == "diabetic retinopathy" or key == "proliferative diabetic retinopathy" or key == "severe nonproliferative retinopathy" or key == "severe proliferative diabetic retinopathy":
            return 1
    
    return 0

def classifica_glaucoma(keywords):
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        if key == "glaucoma" or key == "suspected glaucoma":
            return 1
    
    return 0

def classifica_cataract(keywords):
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        if key == "cataract":
            return 1
    
    return 0

def classifica_degeneracao_macular(keywords):
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        if key == "dry age-related macular degeneration" or  key == "wet age-related macular degeneration" or key == "age-related macular degeneration":
            return 1
    
    return 0

def  classifica_pressao_alta(keywords):
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        if key == "hypertensive retinopathy":
            return 1
    
    return 0

def  classifica_miopia_patologica(keywords):
    keys = keywords.split(separador)
    
    for key in keys:
        key = key.strip()
        if key == "pathological myopia":
            return 1
    
    return 0

def classifica_outras(keywords):
    """
        N - Normal
        D - Diabetes
        G - Glaucoma
        C - Catarata
        A - Degeneração macular
        H - Pressão alta
        M - Miopia patológica 
        O - Outros
    """
    
    
    keys = keywords.split(separador)
    for key in keys:
        key = key.strip()
        if key == "normal fundus":
            continue
    
        if key == "moderate non proliferative retinopathy" or key == "moderate nonproliferative retinopathy" or key == "mild nonproliferative retinopathy" or key == "mild non proliferative retinopathy" or key == "diabetic retinopathy" or key == "proliferative diabetic retinopathy" or key == "severe nonproliferative retinopathy" or key == "severe proliferative diabetic retinopathy":
            continue

        if key == "glaucoma" or key == "suspected glaucoma":
            continue

        if key == "cataract":
            continue

        if key == "dry age-related macular degeneration" or  key == "wet age-related macular degeneration" or key == "age-related macular degeneration":
            continue

        if key == "hypertensive retinopathy":
            continue

        if key == "pathological myopia":
            continue
        
        if key == "anterior segment image" or key == "no fundus image" or key == "lens dust" or key == "optic disk photographically invisible" or key ==  "low image quality":
            continue
        
        return 1
    
    return 0

def class_ds(ds):
    
    ds["keywords"] = ds["keywords"].map(ajusta_separador)

    ds["N"] = ds["keywords"].map(classifica_normal)
    ds["D"] = ds["keywords"].map(classifica_diabetes)
    ds["G"] = ds["keywords"].map(classifica_glaucoma)
    ds["C"] = ds["keywords"].map(classifica_cataract)
    ds["A"] = ds["keywords"].map(classifica_degeneracao_macular)
    ds["H"] = ds["keywords"].map(classifica_pressao_alta)
    ds["M"] = ds["keywords"].map(classifica_miopia_patologica)
    ds["O"] = ds["keywords"].map(classifica_outras)
    ds["X"] = ds["keywords"].map(classifica_desconsiderar)
    
    #return ds;

class_ds(ds_training_eye)
class_ds(ds_validation_eye)
class_ds(ds_test_eye)

#class_ds(ds_all_eye)


ds_training_eye.head()

print(ds_training_eye.query("N == 1").file_name.count())
print(ds_training_eye.query("D == 1").file_name.count())
print(ds_training_eye.query("G == 1").file_name.count())
print(ds_training_eye.query("C == 1").file_name.count())
print(ds_training_eye.query("A == 1").file_name.count())
print(ds_training_eye.query("H == 1").file_name.count())
print(ds_training_eye.query("M == 1").file_name.count())
print(ds_training_eye.query("O == 1").file_name.count())



ds_training_eye["total"]  = ds_training_eye.N +  ds_training_eye.D + ds_training_eye.G + ds_training_eye.C + ds_training_eye.A + ds_training_eye.H + ds_training_eye.M + ds_training_eye.O + ds_training_eye.X
ds_validation_eye["total"]  = ds_validation_eye.N +  ds_validation_eye.D + ds_validation_eye.G + ds_validation_eye.C + ds_validation_eye.A + ds_validation_eye.H + ds_validation_eye.M + ds_validation_eye.O + ds_validation_eye.X
ds_test_eye["total"]  = ds_test_eye.N +  ds_test_eye.D + ds_test_eye.G + ds_test_eye.C + ds_test_eye.A + ds_test_eye.H + ds_test_eye.M + ds_test_eye.O + ds_test_eye.X

#ds_all_eye["total"] = ds_all_eye.N +  ds_all_eye.D + ds_all_eye.G + ds_all_eye.C + ds_all_eye.A + ds_all_eye.H + ds_all_eye.M + ds_all_eye.O + ds_all_eye.X

ds_training_eye



ds_training_eye.query("O == 1 and total == 1")


ds_training_eye.query("file_name == '1467_left.jpg'")

ds_training_eye.query("O == 1")

### Imagens abaixo serão desconsideradas
<p>Todas as imagens classificadas com mais de uma doença será removida do dataset</p>

print("Dataset treinamento")
print("Total com mais de 1 diagnostico:", ds_training_eye.query("total > 1").side.count())
print(ds_training_eye.query("total > 1 and D == 1").side.count())
print(ds_training_eye.query("total > 1 and G == 1").side.count())
print(ds_training_eye.query("total > 1 and C == 1").side.count())
print(ds_training_eye.query("total > 1 and A == 1").side.count())
print(ds_training_eye.query("total > 1 and H == 1").side.count())
print(ds_training_eye.query("total > 1 and M == 1").side.count())
print(ds_training_eye.query("total > 1 and O == 1").side.count())
print(ds_training_eye.query("total > 1 and X == 1").side.count())

print("Dataset validação")
print("Total com mais de 1 diagnostico:", ds_training_eye.query("total > 1").side.count())
print(ds_validation_eye.query("total > 1 and D == 1").side.count())
print(ds_validation_eye.query("total > 1 and G == 1").side.count())
print(ds_validation_eye.query("total > 1 and C == 1").side.count())
print(ds_validation_eye.query("total > 1 and A == 1").side.count())
print(ds_validation_eye.query("total > 1 and H == 1").side.count())
print(ds_validation_eye.query("total > 1 and M == 1").side.count())
print(ds_validation_eye.query("total > 1 and O == 1").side.count())
print(ds_validation_eye.query("total > 1 and X == 1").side.count())

print("Dataset teste")
print("Total com mais de 1 diagnostico:", ds_training_eye.query("total > 1").side.count())
print(ds_test_eye.query("total > 1 and D == 1").side.count())
print(ds_test_eye.query("total > 1 and G == 1").side.count())
print(ds_test_eye.query("total > 1 and C == 1").side.count())
print(ds_test_eye.query("total > 1 and A == 1").side.count())
print(ds_test_eye.query("total > 1 and H == 1").side.count())
print(ds_test_eye.query("total > 1 and M == 1").side.count())
print(ds_test_eye.query("total > 1 and O == 1").side.count())
print(ds_test_eye.query("total > 1 and X == 1").side.count())



### Imagens com apenas 1 classificação
<p>Essa imagens serão utilizadas para o treinamento</p>

print("Dataset treinamento")
print("Total com apenas 1 diagnostico:", ds_training_eye.query("total == 1").side.count())
print(ds_training_eye.query("total == 1 and N == 1").side.count())
print(ds_training_eye.query("total == 1 and D == 1").side.count())
print(ds_training_eye.query("total == 1 and G == 1").side.count())
print(ds_training_eye.query("total == 1 and C == 1").side.count())
print(ds_training_eye.query("total == 1 and A == 1").side.count())
print(ds_training_eye.query("total == 1 and H == 1").side.count())
print(ds_training_eye.query("total == 1 and M == 1").side.count())
print(ds_training_eye.query("total == 1 and O == 1").side.count())
print(ds_training_eye.query("total == 1 and X == 1").side.count())

print("Dataset validação")
print("Total com apenas 1 diagnostico:", ds_training_eye.query("total == 1").side.count())
print(ds_validation_eye.query("total == 1 and N == 1").side.count())
print(ds_validation_eye.query("total == 1 and D == 1").side.count())
print(ds_validation_eye.query("total == 1 and G == 1").side.count())
print(ds_validation_eye.query("total == 1 and C == 1").side.count())
print(ds_validation_eye.query("total == 1 and A == 1").side.count())
print(ds_validation_eye.query("total == 1 and H == 1").side.count())
print(ds_validation_eye.query("total == 1 and M == 1").side.count())
print(ds_validation_eye.query("total == 1 and O == 1").side.count())
print(ds_validation_eye.query("total == 1 and X == 1").side.count())

print("Dataset teste")
print("Totalcom apenas 1 diagnostico:", ds_training_eye.query("total == 1").side.count())
print(ds_test_eye.query("total == 1 and N == 1").side.count())
print(ds_test_eye.query("total == 1 and D == 1").side.count())
print(ds_test_eye.query("total == 1 and G == 1").side.count())
print(ds_test_eye.query("total == 1 and C == 1").side.count())
print(ds_test_eye.query("total == 1 and A == 1").side.count())
print(ds_test_eye.query("total == 1 and H == 1").side.count())
print(ds_test_eye.query("total == 1 and M == 1").side.count())
print(ds_test_eye.query("total == 1 and O == 1").side.count())
print(ds_test_eye.query("total == 1 and X == 1").side.count())


ds_training_eye.query("total == 1 and N == 1")

ds_test_eye.query("total == 1 and N == 1")

### Criando um novo data
<p> Aqui vamos criar um dataset com imagens apenas 1 classificação e vamos descartar as doenças abaixo:</p>
    <ul>
    <li>Glaucoma - Não tem imagens suficientes para um bom treinamento</li>
    <li>Pressão alta - Precimos de imagens com definição melhor</li>
    <li>Outros - São diversar doenças com caracteristicas distintas</li>
    </ul>
    
    

ds_new_train = ds_training_eye.query("total == 1  and O == 0 and X == 0")
ds_new_valid = ds_validation_eye.query("total == 1 and O == 0 and X == 0")
ds_new_test = ds_test_eye.query("total == 1  and O == 0 and X == 0")

ds_new_train.head(15)

ds_new_valid

ds_new_test

### Salvando os novos dataset

ds_new_train.to_excel("training_final.xlsx") 
ds_new_valid.to_excel("validation_final.xlsx") 
ds_new_test.to_excel("test_final.xlsx")
