==> 03.2-Pre_processamento - Obtendo a media e desvioPadrao.ipynb <==
# Obtendo o média e o desvio padrão para a execução
<p> Para realizar o treinamento é necessário normalizar as imagens, para isso precisamos encontrar a Média e o Desvio padrão.</p>

from pathlib import Path
import numpy as np
import pandas as pd
import torch

import matplotlib.pyplot as plt
import cv2

from tqdm.notebook import tqdm

import torchvision
import torchvision.transforms as transforms
from torch.utils.data import DataLoader

train_transforms = transforms.Compose([
    transforms.ToTensor(),
    
    ])


#Pasta para salvar o Dataset para treinamento e validação
PATH_DS_PROCESSED_TRAINING = Path("./IMG224_ALL/PROCESSED-TRAIN/")
PATH_DS_PROCESSED_VALIDATION = Path(".IMG224_ALL/PROCESSED-VALID/")
PATH_DS_PROCESSED_TEST = Path("./IMG224_ALL/PROCESSED-TEST/")

def load_file(path):
    return np.load(path).astype(np.float32)

TRAIN_DATASET = torchvision.datasets.DatasetFolder(
    PATH_DS_PROCESSED_TEST,
    loader=load_file, extensions="npy", transform=transforms.ToTensor())

# python code to calculate mean and std 

loader = DataLoader(
  TRAIN_DATASET, 
  batch_size = 4, 
  num_workers=0)

def batch_mean_and_sd(loader):
    
    cnt = 0
    fst_moment = torch.empty(3)
    snd_moment = torch.empty(3)

    for images, _ in loader:
        b, c, h, w = images.shape
        nb_pixels = b * h * w
        sum_ = torch.sum(images, dim=[0, 2, 3])
        sum_of_square = torch.sum(images ** 2,
                                  dim=[0, 2, 3])
        fst_moment = (cnt * fst_moment + sum_) / (
                      cnt + nb_pixels)
        snd_moment = (cnt * snd_moment + sum_of_square) / (
                            cnt + nb_pixels)
        cnt += nb_pixels

    mean, std = fst_moment, torch.sqrt(
      snd_moment - fst_moment ** 2)        
    return mean,std
  
mean, std = batch_mean_and_sd(loader)
print("mean and std: \n", mean, std)
